Move data classes into model package
