+++ domain/widget/src/commonMain/kotlin/com/thomaskioko/tvmaniac/domain/widget/WidgetPublisher.kt
@@ -1,5 +1,7 @@
 package com.thomaskioko.tvmaniac.domain.widget
 
+import com.thomaskioko.tvmaniac.domain.widget.model.WidgetShow
+
 public interface WidgetPublisher {
 
     public suspend fun hasInstalledWidgets(): Boolean
